apps/api/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
pg8000==1.29.3
redis==4.6.0
rq==1.12.0
python-multipart==0.0.6
minio==7.1.2
prometheus_client==0.18.0
alembic==1.11.1

# Auth0/OIDC support (optional - for production auth)
authlib==1.2.1
python-jose[cryptography]==3.3.0
httpx==0.25.0
itsdangerous==2.1.2
